--- LifeMatch/CRUD.sql
@@ -1,5 +1,0 @@
-show tables;
-select*from tb_postagem;
-select*from tb_tema;
-select*from tb_usuario;
-
