--- LifeMatch/LifeMatch/src/main/java/com/Ampara/LifeMatch/ProjetoIntegradorApplication.java
@@ -1,13 +1,0 @@
-package com.Ampara.LifeMatch;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProjetoIntegradorApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ProjetoIntegradorApplication.class, args);
-	}
-
-}
