--- LifeMatch/LifeMatch/src/main/java/com/Ampara/LifeMatch/controller/PostagemController.java
@@ -1,61 +1,0 @@
-package com.Ampara.LifeMatch.controller;
-
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.Ampara.LifeMatch.model.PostagemModel;
-import com.Ampara.LifeMatch.repository.PostagemRepository;
-
-@RestController
-@RequestMapping ("/postagem")
-@CrossOrigin(origins = "*", allowedHeaders = "*")
-public class PostagemController {
-	@Autowired
-	private PostagemRepository repository;
-	
-	@GetMapping
-	public ResponseEntity<List<PostagemModel>>getAll(){
-		return ResponseEntity.ok(repository.findAll());
-	}
-	
-	@GetMapping("/{id}")
-	public ResponseEntity<PostagemModel>getById(@PathVariable Long id){
-		return repository.findById(id)
-				.map(resp-> ResponseEntity.ok(resp))
-						.orElse(ResponseEntity.notFound().build());
-	}
-	
-	@GetMapping("/{titulo}")
-	public ResponseEntity<PostagemModel>getByCategoriaAjuda(@PathVariable String titulo){
-		return repository.findByTitulo(titulo)
-				.map(resp-> ResponseEntity.ok(resp))
-						.orElse(ResponseEntity.notFound().build());
-	}
-	
-	@PostMapping
-	public ResponseEntity<PostagemModel>post (@RequestBody PostagemModel postagem){
-		return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
-	}
-	
-	@PutMapping
-	public ResponseEntity<PostagemModel>put (@RequestBody PostagemModel postagem){
-		return ResponseEntity.ok(repository.save(postagem));
-	}
-	
-	@DeleteMapping("/{id}")
-	public void delete(@PathVariable Long id) {
-		repository.deleteById(id);
-	}
-}
