--- LifeMatch/LifeMatch/src/main/java/com/Ampara/LifeMatch/controller/TemaController.java
@@ -1,64 +1,0 @@
-package com.Ampara.LifeMatch.controller;
-
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.Ampara.LifeMatch.model.TemaModel;
-import com.Ampara.LifeMatch.repository.TemaRepository;
-
-@RestController
-@RequestMapping ("/tema")
-@CrossOrigin(origins = "*", allowedHeaders = "*")
-public class TemaController {
-	@Autowired
-	private TemaRepository repository;
-	
-	@GetMapping
-	public ResponseEntity<List<TemaModel>>getAll(){
-		return ResponseEntity.ok(repository.findAll());
-	}
-	
-	@GetMapping("/{id}")
-	public ResponseEntity<TemaModel>getById(@PathVariable Long id){
-		return repository.findById(id)
-				.map(resp-> ResponseEntity.ok(resp))
-						.orElse(ResponseEntity.notFound().build());
-	}
-	
-	@GetMapping("/{categoriaAjuda}")
-	public ResponseEntity<TemaModel>getByCategoriaAjuda(@PathVariable String categoriaAjuda){
-		return repository.findByCategoriaAjuda(categoriaAjuda)
-				.map(resp-> ResponseEntity.ok(resp))
-						.orElse(ResponseEntity.notFound().build());
-	}
-	
-	
-	
-	@PostMapping
-	public ResponseEntity<TemaModel>post (@RequestBody TemaModel tema){
-		return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema));
-	}
-	
-	@PutMapping
-	public ResponseEntity<TemaModel>put (@RequestBody TemaModel tema){
-		return ResponseEntity.ok(repository.save(tema));
-	}
-	
-	@DeleteMapping("/{id}")
-	public void delete(@PathVariable Long id) {
-		repository.deleteById(id);
-	}
-	
-}
