--- LifeMatch/LifeMatch/src/main/java/com/Ampara/LifeMatch/controller/UsuarioController.java
@@ -1,63 +1,0 @@
-package com.Ampara.LifeMatch.controller;
-
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.Ampara.LifeMatch.model.PostagemModel;
-import com.Ampara.LifeMatch.model.UsuarioModel;
-import com.Ampara.LifeMatch.repository.UsuarioRepository;
-
-@RestController
-@RequestMapping ("/usuario")
-@CrossOrigin(origins = "*", allowedHeaders = "*")
-public class UsuarioController {
-	@Autowired
-	private UsuarioRepository repository;
-	
-	@GetMapping
-	public ResponseEntity<List<UsuarioModel>>getAll(){
-		return ResponseEntity.ok(repository.findAll());
-	}
-	
-	@GetMapping("/{id}")
-	public ResponseEntity<UsuarioModel>getById(@PathVariable Long id){
-		return repository.findById(id)
-				.map(resp-> ResponseEntity.ok(resp))
-						.orElse(ResponseEntity.notFound().build());
-	}
-	
-	@GetMapping("/{loginUsuario}")
-	public ResponseEntity<UsuarioModel>getByCategoriaAjuda(@PathVariable String loginUsuario){
-		return repository.findByLoginUsuario(loginUsuario)
-				.map(resp-> ResponseEntity.ok(resp))
-						.orElse(ResponseEntity.notFound().build());
-	}
-	
-	@PostMapping
-	public ResponseEntity<UsuarioModel>post (@RequestBody UsuarioModel usuario){
-		return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(usuario));
-	}
-	
-	@PutMapping
-	public ResponseEntity<UsuarioModel>put (@RequestBody UsuarioModel usuario){
-		return ResponseEntity.ok(repository.save(usuario));
-	}
-	
-	@DeleteMapping("/{id}")
-	public void delete(@PathVariable Long id) {
-		repository.deleteById(id);
-	}
-	
-}
