--- LifeMatch/LifeMatch/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.properties.hibernate.format_sql=true
-spring.datasource.url=jdbc:mysql://localhost/db_life_match?createDatabaseIfNotExist=true&serverTimezone=UTC&useSSl=false
-spring.datasource.username=root
-spring.datasource.password=
-spring.jpa.show-sql=true
-
-
