--- LifeMatch/LifeMatch/src/test/java/com/Ampara/LifeMatch/ProjetoIntegradorApplicationTests.java
@@ -1,13 +1,0 @@
-package com.Ampara.LifeMatch;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProjetoIntegradorApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
