--- LifeMatch/Teste.txt
@@ -1,1 +1,0 @@
-sasa
