--- LifeMatch/src/main/java/com/Ampara/LifeMatch/Enums/CategoriaAjuda.java
@@ -1,8 +1,0 @@
-package com.Ampara.LifeMatch.Enums;
-
-
-public enum CategoriaAjuda {
-	Higiene,
-	Alimenticia,
-	Profissional
-}
