--- LifeMatch/src/main/java/com/Ampara/LifeMatch/Enums/Identificacao.java
@@ -1,6 +1,0 @@
-package com.Ampara.LifeMatch.Enums;
-
-public enum Identificacao {
-		ANONIMO,
-		PUBLICO
-}
