+++ LifeMatch/src/main/java/com/Ampara/LifeMatch/model/TemaModel.java
@@ -1,5 +1,8 @@
 package com.Ampara.LifeMatch.model;
 
+import java.util.List;
+
+import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
@@ -7,10 +10,11 @@ import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.validation.constraints.Size;
 
-import com.Ampara.LifeMatch.Enums.CategoriaAjuda;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.sun.istack.NotNull;
 
 @Entity
@@ -19,12 +23,12 @@ public class TemaModel {
 	
 	@Id
 	@GeneratedValue (strategy = GenerationType.IDENTITY)
-	private Long id_tema;
+	private Long idTema;
 	
 	
-	@Column(columnDefinition = "ENUM('Higiene', 'Alimenticia', 'Profissional')")
-    @Enumerated(EnumType.STRING)
-    private CategoriaAjuda categoriaAjuda ;
+	@Column
+    @NotNull
+    private String categoriaAjuda ;
 
 	
 	@Column
@@ -32,25 +36,42 @@ public class TemaModel {
 	@Size (min = 10 ,max = 500)
 	private String decricao;
 
+	@OneToMany (mappedBy = "tema", cascade = CascadeType.ALL)
+	@JsonIgnoreProperties ("tema")
+	private  List<PostagemModel> postagem;
 
-	public Long getId_tema() {
-		return id_tema;
+	public Long getIdTema() {
+		return idTema;
 	}
 
-
-	public void setId_tema(Long id_tema) {
-		this.id_tema = id_tema;
+	public void setIdTema(Long idTema) {
+		this.idTema = idTema;
 	}
 
+	public String getCategoriaAjuda() {
+		return categoriaAjuda;
+	}
+
+	public void setCategoriaAjuda(String categoriaAjuda) {
+		this.categoriaAjuda = categoriaAjuda;
+	}
 
 	public String getDecricao() {
 		return decricao;
 	}
 
-
 	public void setDecricao(String decricao) {
 		this.decricao = decricao;
 	}
-		
+
+	public List<PostagemModel> getPostagem() {
+		return postagem;
+	}
+
+	public void setPostagem(List<PostagemModel> postagem) {
+		this.postagem = postagem;
+	}
+	
+	
 	
 }
