+++ LifeMatch/src/main/java/com/Ampara/LifeMatch/repository/UsuarioRepository.java
@@ -1,10 +1,15 @@
 package com.Ampara.LifeMatch.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.Ampara.LifeMatch.model.UsuarioModel;
+
 @Repository
 public interface UsuarioRepository extends JpaRepository<UsuarioModel, Long> {
+
+	Optional<UsuarioModel> findByLoginUsuario(String loginUsuario);
 
 }
